refactor(bulk-session): 조합키 해석기를 공용 모듈로 뽑는다
발행 경로가 같은 해석을 써야 한다 — 복제하면 수정 행(idKey)과 신규 행
(이름 기반)의 비대칭이 두 벌이 되어 한쪽만 고쳐진다.

순수 이동이다. 기존 스펙을 고치지 않고 통과해야 한다.
